Update crux.sh
Fixed Step 1.1 to find Obitools databases. The script looks for any name that has a file ending in .tdx. Each packet of data from EMBL that has been converted to Obitools format will have multiple .sdx files but only a single .adx, .rdx, .ndx, and .tdx, so we leverage this to get the unique name for the entire bunch of files. Thanks to Nathan Schaefer in the Shapiro/Green lab for rewriting the python!